#25: Limited notification management to admins.

+++ app/Commands/SaveAccountSettingsCommand.php
@@ -24,7 +24,6 @@ use DB;
 use Poniverse\Ponyfm\Models\Image;
 use Poniverse\Ponyfm\Models\User;
 use Gate;
-use Auth;
 use Validator;
 
 class SaveAccountSettingsCommand extends CommandBase
@@ -111,20 +110,23 @@ class SaveAccountSettingsCommand extends CommandBase
             $this->_user->save();
 
             // Sync email subscriptions
-            $emailSubscriptions = $this->_user->emailSubscriptions->keyBy('activity_type');
-            foreach ($this->_input['notifications'] as $notificationSetting) {
+            // TODO: [#25] Remove this when email notifications are rolled out to everyone.
+            if (Gate::forUser($this->_user)->allows('receive-email-notifications')) {
+                $emailSubscriptions = $this->_user->emailSubscriptions->keyBy('activity_type');
+                foreach ($this->_input['notifications'] as $notificationSetting) {
 
-                if (
-                    $notificationSetting['receive_emails'] &&
-                    !$emailSubscriptions->offsetExists($notificationSetting['activity_type'])
-                ) {
-                    $this->_user->emailSubscriptions()->create(['activity_type' => $notificationSetting['activity_type']]);
+                    if (
+                        $notificationSetting['receive_emails'] &&
+                        !$emailSubscriptions->offsetExists($notificationSetting['activity_type'])
+                    ) {
+                        $this->_user->emailSubscriptions()->create(['activity_type' => $notificationSetting['activity_type']]);
 
-                } elseif (
-                    !$notificationSetting['receive_emails'] &&
-                    $emailSubscriptions->offsetExists($notificationSetting['activity_type'])
-                ) {
-                    $emailSubscriptions->get($notificationSetting['activity_type'])->delete();
+                    } elseif (
+                        !$notificationSetting['receive_emails'] &&
+                        $emailSubscriptions->offsetExists($notificationSetting['activity_type'])
+                    ) {
+                        $emailSubscriptions->get($notificationSetting['activity_type'])->delete();
+                    }
                 }
             }
         });
